Update .gitignore (for non-forked repo)
test/gtest is very important for the compilation process. If students copy the files into another repo, gtest folder would not be pushed into the new repo (if students decide not to fork). Then, they may not copy the correct files on to the Pi for it to successfully compile.